capstone_deployment/core/openai_generators/summarizer_fixer.py
from openai import OpenAI
from dotenv import load_dotenv
import os
import json

load_dotenv()

client = OpenAI(api_key=os.environ.get('OPENAI_API_KEY'))

def fix_summary(paragraph: str):
    prompt = (
        '''
    You are an experienced tutor preparing a student for an important exam. Your task is to create a detailed, well-organized summary notes that covers the whole provided material. The summary should be:

            - Concise (Eliminate repetitive points)
            - Objective (Avoid introducing personal biases, and stick to the facts as presented in the original content)
            - Clarity (Structuring sentences in a straightforward manner)
            - Coherence (Maintain a logical sequence, often following the structure of the original content)
            - Lastly, use Extractive Summary (do not rephrase) especially in important concepts, and terms and definitions.
            - You may use Abstractive Summary (rephrasing words using Layman's Terms) in Introduction and Overview Summary only.

            The summary notes should follow this structure:
            - **Introduction**
            - **Multiple Key Concepts**
            - **Overview Summary**

            ### Guidelines for Introduction:
            - Should be 1 paragraph only containing 3 sentences (minimum) to 5 sentences (maximum) long.
            - Begin with an engaging hook, such as a thought-provoking question, a surprising fact, or a relevant quote to capture the reader's attention.
            - Clearly state the objective of the summary notes.
            - Should highlight the key areas that the summary notes will cover.
            - Mention the importance to the reader.
            - Avoid starting with phrases like "This summary note..." or "In this guide...".

            ### Guidelines for Key Concepts:
            - Core Ideas
            - Terms and Definitions (Include even the sub-bullets or sub-points)
            - Provide only one that fits each core idea, either (Real-Life Application or Analogy) to better understand the Key Concepts

            ### Guidelines for Overview Summary:
            - Should be one paragraph only containing 3 sentences (minimum) to 5 sentences (maximum) long.
            - Summarize the main points in a compelling and creative manner.
            - Provide final thoughts that encapsulate the entire summary note.
            - End with a forward-looking statement or a call to action to encourage further exploration of the topic.
            - Avoid starting with phrases like "This summary note..." or "In conclusion...".

            ### Important Instruction:
            - **RESPOND ONLY IN HTML FORMAT** but **DO NOT include** the `<html>`, `<head>`, `<title>`, or `<body>` tags. Use appropriate HTML tags to structure the content with headings, lists, and paragraphs. Maintain the logical sequence and structure provided, and ensure each section is clearly defined.

            Here is the example format of the response:

            <h2>Introduction</h2>
            <p>In an era where cyber threats loom larger than ever, mastering security and access control is not just beneficial—it's essential. Delving into principles like least privilege and separation of duties empowers you to fortify systems against vulnerabilities. This guide illuminates these critical concepts to enhance your proficiency in safeguarding digital assets.</p>

            <h2>Key Concepts</h2>
            <h3>1. Disaster Recovery as a Service (DRaaS)</h3>
            <ul>
                <li>This is where a third-party provider offers disaster recovery solutions via the cloud. </li>
            </ul>
            <h3>2. Principles of Security and Access Control</h3>
            <ul>
            <li><strong>Least Privilege</strong>: Users are given only the minimum access necessary to perform their tasks, reducing the risk of damage.</li>
            <li><strong>Separation of Duties</strong>: Different individuals perform different administrative tasks to prevent conflicts of interest and abuse.</li>
            </ul>
            <p><em>Real-Life Application</em>: Like restricting access to sensitive documents to only relevant employees to avoid data leaks.</p>

            <h3>3. Incident Response and Recovery</h3>
            <ul>
            <li><strong>Incident Response Plan</strong>: A structured approach to handling security breaches, including:
                <ul>
                <li><strong>Identification</strong>: Recognizing the incident.</li>
                <li><strong>Containment</strong>: Minimizing the damage.</li>
                </ul>
            </li>
            <li><strong>Backup and Recovery</strong>: Regular data backups and recovery procedures ensure systems can be restored in case of failure or data loss.</li>
            </ul>
            <p><em>Analogy</em>: An incident response plan is like a fire drill plan—clear roles and actions minimize harm during an emergency.</p>

            <h2>Overview Summary</h2>
            <p>Embracing these security principles equips you to navigate the complex landscape of system administration confidently. By implementing robust access controls and being prepared for incidents, you safeguard your organization's vital information. Continue to build on this knowledge to stay ahead in the ever-evolving field of cybersecurity.</p>

        '''
    )

    response = client.chat.completions.create(
        model='gpt-4o',
        messages=[
            {"role": "system", "content": prompt },
            {"role": "user", "content": paragraph},
        ],
        temperature=0.7,
    )

    title = response.choices[0].message.content


    return title
