capstone_deployment/core/serializers/badges.py
from rest_framework import serializers
from core.models.badges import Badge, Achievement

class BadgeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Badge
        fields = "__all__"

class AchievementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Achievement
        fields = "__all__"
