capstone_deployment/requirements/base.txt
acme==2.9.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
certbot==2.9.0
certbot-nginx==2.9.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.6
colorama==0.4.6
ConfigArgParse==1.7
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
defusedxml==0.7.1
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
oauthlib==3.2.2
openai==1.65.4
packaging==24.0
parsedatetime==2.6
pexpect==4.9.0
psycopg==3.2.5
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
pywin32==308
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.0.7
wadllib==1.3.6
whitenoise==6.9.0
zope.interface==6.1